requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
blinker==1.9.0
    # via flask
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via flask
flask==3.1.0
    # via -r requirements.in
idna==3.10
    # via requests
iniconfig==2.1.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
packaging==24.2
    # via pytest
pluggy==1.5.0
    # via pytest
pytest==8.3.5
    # via -r requirements.in
python-dotenv==1.1.0
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
urllib3==2.3.0
    # via requests
werkzeug==3.1.3
    # via flask
